LWDEV-6992 fees use big decimal

+++ src/main/kotlin/com/lykke/matching/engine/daos/FeeInstruction.kt
@@ -3,11 +3,12 @@ package com.lykke.matching.engine.daos
 import com.lykke.matching.engine.daos.fee.NewFeeInstruction
 import com.lykke.matching.engine.messages.ProtocolMessages
 import java.io.Serializable
+import java.math.BigDecimal
 
 open class FeeInstruction(
         val type: FeeType,
         val sizeType: FeeSizeType?,
-        val size: Double?,
+        val size: BigDecimal?,
         val sourceClientId: String?,
         val targetClientId: String?
 ) : Serializable {
@@ -25,7 +26,7 @@ open class FeeInstruction(
             return FeeInstruction(
                     feeType,
                     sizeType,
-                    if (fee.hasSize()) fee.size else null,
+                    if (fee.hasSize()) BigDecimal.valueOf(fee.size)  else null,
                     if (fee.hasSourceClientId()) fee.sourceClientId else null,
                     if (fee.hasTargetClientId()) fee.targetClientId else null
             )
